requirements.txt
apex
braceexpand
contextlib
einops
einops-exts
fairseq==0.12.2
faiss-cpu==1.7.3
ftfy
h5py
huggingface-hub
laion-clap==1.1.3
librosa==0.8.0
llvmlite==0.36.0
numba
numpy
omegaconf==2.0.6
pandas
pillow
progress==1.6
progressbar==2.5
pydub
pyyaml
regex
scikit-learn
scipy
sentence-transformers==2.2.2
sentencepiece==0.1.97
soundfile
sox==1.4.1
tensorboard
torch==2.0.1
torchaudio==2.0.2
torchlibrosa==0.1.0
torchvision==0.15.2
tqdm
transformers==4.27.4
webrtcvad==2.0.10
